Potential bug with Kd update
It seems that on loop #0, it will do the if statement. Then it will bump _is_first_loop to 1. On loop #1, it will go to the else statement. Here, it will use (for Kd): err - prev_error / dt
But in this case, prev_err is not defined until the 3rd loop, when you set it equal to err.